refactor(types): extract artifact chat note and sharing models
